refactor: 改名，宏降低冗余
