+++ src/policy/usage_top2/policy_party.rs
@@ -1,5 +1,2 @@
-use crate::policy::pkg_cfg::StartArgs;
-
-pub fn start_task(args: &mut StartArgs<'_>) {
-    super::StartTask::new(args).start_task(b"MainThread", Some(b"Thread-"));
-}
+use super::top2_macro_init;
+top2_macro_init!(b"MainThread", Some(b"Thread-"));
